feat: refactor BottomToolBar into segments; implement separate components for left, middle, and right sections
