+++ src/context/TrainingContext.tsx
@@ -26,6 +26,10 @@ export const TrainingProvider = ({ children }: { children: ReactNode }) => {
   const addEdgeToUserPath = (edgeId: string, isReverse: boolean) => {
     if (!edgeId) return;
     setUserPath((prev) => [...prev, edgeId]);
+    console.log(
+      `Edge ${edgeId} ${isReverse ? "reversed" : "added"} to user path`,
+      userPath,
+    );
   };
 
   const clearUserPath = () => {
